refactor: 훅으로 로직 분리 및 sellerId로 변경
